requirements.txt
backports.weakref==1.0.post1
bleach==1.5.0
certifi==2016.2.28
cmake==3.18.2.post1
distro==1.5.0
enum34==1.1.10
funcsigs==1.0.2
futures==3.3.0
html5lib==0.9999999
Markdown==3.1.1
matlabengineforpython===R2017a
mock==3.0.5
numpy==1.16.6
opencv-python==3.2.0.8
packaging==20.4
Pillow==6.2.2
protobuf==3.13.0
pyparsing==2.4.7
scikit-build==0.11.1
scikit-learn==0.20.4
scipy==1.2.3
six==1.15.0
sklearn==0.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0
Werkzeug==1.0.1

